Expose interface for DownloadMigrator, hiding the class.
