+++ library/src/main/java/com/novoda/downloadmanager/DownloadMigratorBuilder.java
@@ -40,6 +40,6 @@ public class DownloadMigratorBuilder {
     }
 
     public DownloadMigrator build() {
-        return new DownloadMigrator(applicationContext, handler, notificationChannelCreator, notificationCreator);
+        return new LiteDownloadMigrator(applicationContext, handler, notificationChannelCreator, notificationCreator);
     }
 }
